refactor: DBConfigLoader -> ConfigLoader 변경
